adgm_corporate_agent/requirements.txt
python-docx
PyPDF2
requests
sentence-transformers
faiss-cpu
chromadb
gradio
openai
pdfplumber
numpy
